+++ src/main/java/com/codecool/hccrm/repository/implementation/CompanyQueryRepositoryImpl.java
@@ -1,7 +1,8 @@
-package com.codecool.hccrm.repository;
+package com.codecool.hccrm.repository.implementation;
 
 import com.codecool.hccrm.model.Company;
 import com.codecool.hccrm.model.Condominium;
+import com.codecool.hccrm.repository.CompanyQueryRepository;
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
 import org.springframework.beans.factory.annotation.Autowired;
